+++ test/DependencyInjection/Loader/YamlFileLoader.spec.js
@@ -30,19 +30,25 @@ describe('YamlFileLoader', () => {
       let actual = () => loader.load()
 
       // Assert.
-      assert.throws(actual, Error, 'The file ' + path + ' not exists')
+      return assert.throws(actual, Error, 'The file ' + path + ' not exists')
     })
 
     it('should load a simple container', () => {
       // Arrange.
       let serviceName = 'foo'
       let aliasName = 'f'
+      let tagName = 'fooTag'
+      let stringParameterName = 'fooParameter'
+      let stringExpectedParameter = 'barValue'
+      let arrayParameterName = 'barParameter'
 
       // Act.
       loader.load()
       let service = container.get(serviceName)
       let aliasService = container.get(aliasName)
-      let taggedServices = container.findTaggedServiceIds('fooTag')
+      let taggedServices = container.findTaggedServiceIds(tagName)
+      let stringActualParameter = container.getParameter(stringParameterName)
+      let arrayActualParameter = container.getParameter(arrayParameterName)
 
       // Assert.
       assert.instanceOf(service, Foo)
@@ -52,6 +58,11 @@ describe('YamlFileLoader', () => {
       assert.strictEqual(service.param, 'foo-bar')
       assert.strictEqual(aliasService, service)
       assert.lengthOf(taggedServices.toArray(), 2)
+      assert.strictEqual(stringActualParameter, stringExpectedParameter)
+      assert.isArray(arrayActualParameter)
+      assert.strictEqual(service.parameter, stringExpectedParameter)
+
+      return assert.lengthOf(arrayActualParameter, 2)
     })
   })
 
@@ -73,7 +84,8 @@ describe('YamlFileLoader', () => {
 
       // Assert.
       assert.instanceOf(service1, Foo)
-      assert.instanceOf(service2, FooManager)
+
+      return assert.instanceOf(service2, FooManager)
     })
   })
 })